MOD-39 - flush out use of tempnam for generating temp key files.
Rename getBaseDir -> getTmpDir
Create a protected wrapper method around tempnam, allowing the method to be mocked and filesystem access averted.
Move the initial generation of the temp files out of the constructor. Generate the files the first time they are requested via getPublicKeyPath and getPrivateKeyPath. Allows for the call to _tempnam to be mocked.
Change private methods to protected so they can be mocked in other tests.
Update the SFTP protocol model to no longer explicitly call destroyKeys, this will happen implicitly when the key maker model is dereferenced.
Updates to unit tests.