MOD-56 - Allow for all files matching a pattern to be retrieved from a remote.
Add a getAllFiles method to the sftp model. Method expects to be passed a local target directory, a remote source directory and an optional glob that files must match. Method will get any files in the source directory that match the glob and copy them to the target directory.
Add directory centric methods to the sftp adapter - opendir, readdir and closedir.