US7525 Autonomous FileTransfer
Provide an autonomous mechanism for FileTransfer. This mechanism will read a
set of registered configuration for remote hosts and sets of local and remote
directories to transfer files between.

Create an observer method to run the transfer process. The observer will
gather up configuration for FileTransfer and run through each of the hosts,
importing and exporting all files between the configured directories.

Add all of the necessary methods that a protocol model must implement to
facilitate the autonomous FileTransfer process to the protocol abstract.
This included adding method to get all files in a remote directory and a
method to send all files in a local directory.

Implement all necessary methods in the currently supported SFTP protocol.

Add a filetransfer/file helper for working with local files. Includes methods
to list all files in a local directory matching a pattern and moving a file
from one location to another, creating directories as needed.

Delete method from the filetransfer/data helper for doing one-off transfers.
While still possible to make such requests via the protocol models directly,
it is no longer the preferred mechanism for doing transfers and as such,
means to facilitate such use are no longer necessary.

Reorganize documentation. Root README becomes a high level overview of the
included modules with links to more detailed documentation. Add a `docs`
directory to the root to house the more detailed documentation. This directory
will contain further directories and .md files to organize documentation
specific to a specific module, implementation or other aspect of the project.